feat: FlaSON CLI
Creates the FlaSON CLI to read JSON from stdin and print the RFC 6902
JSON pointer and JSON value for each scalar value. The JSON token
parsing and the RFC 6902 types are defined in a library but the logic to
convert the stream of JSON tokens into FlaSON is currently in the CLI.
Next steps are to produce a better interface for JSON token parsing,
perhaps IAsyncEnumerable, to move the JSON -> FlaSON logic into the
library, and publish the library and CLI to NuGet, GitHub releases, and
a scoop bucket.