q: why doesn't the compiler know this is safe?
The new function borrows if Some(_) is returned and doesn't if None is
returned but the compiler doesn't recognize that the None branch makes
it safe. This is surprising because before it's factored out the same
thing is true and the compiler is fine with it in that case. What is the
important difference and is there a way to make this function work?